requirements.txt
jmespath
molecule
pytest
pytest-cov
ansible-core
